+++ src/components/Main/MainContent.js
@@ -1,19 +1,20 @@
 import { useState } from "react";
 import styled from "styled-components";
+import Category from "../Category/Category";
 import CategoryTag from "../CategoryTag/CategoryTag";
 
-const MainContent = () => {
-  const [categories, setCategories] = useState([]);
+const MainContent = ({categories, setCategories, setShowCategorySetting}) => {
+
   const [selectedCategoryIndex, setSelectedCategoryIndex] = useState(null);
 
   return (
     <MainContentWrapper>
-      <div style={{display: "flex", marginTop: "8px", paddingBottom: "16px"}}>
-        <CategoryTag categories={categories} setCategories={setCategories} selectedCategoryIndex={selectedCategoryIndex} setSelectedCategoryIndex={setSelectedCategoryIndex} />
+      <div style={{display: "flex", paddingBottom: "16px", position: "sticky", top: "0", width: "100%", backgroundColor: "#FFFFFF"}}>
+        <CategoryTag categories={categories} setCategories={setCategories} selectedCategoryIndex={selectedCategoryIndex} setSelectedCategoryIndex={setSelectedCategoryIndex} setShowCategorySetting={setShowCategorySetting} />
       </div>
-      <div>
-        카테고리에 따른 기도제목 목록
-        {categories[selectedCategoryIndex].name}
+      <div style={{display: "flex", flexDirection: "column", gap: "24px"}}>
+        <Category title="테스트" color="#75BD62" />
+        <Category title="테스트2" color="#AEDBA5" />
       </div>
     </MainContentWrapper>
   );
@@ -30,5 +31,16 @@ const MainContentWrapper = styled.div`
   background-color: var(--color-white);
   border-radius: 32px 32px 0px 0px;
   padding: 16px;
+  padding-bottom: 64px;
   box-sizing: border-box;
+  overflow-y: auto;
+
+  // Chrome, Edge 등
+  &::-webkit-scrollbar {
+    display: none;
+  }
+  // Firefox
+  scrollbar-width: none;
+  // IE, Edge
+  -ms-overflow-style: none;
 `;
